Fixed issue in F3 view distance script
For some reason, `f_var_debugmode` was reporting as `undefined` in the
F3 view distance script; I've added a `waitUntil` there which seems to
have fixed the Issue, but I am not sure why this started all of the
sudden. Could be due to reduction in calls to other scripts and in
parameter count making initialization happen faster.